feat(ui): change fullscreen preview to a popup
